% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruteknapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force method for solving the knapsack problem}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{a data frame has two columns and all the elements are positive numbers}

\item{W}{a single number denotes the size of the knapsack}

\item{parallel}{default FALSE: set to TRUE if you wish to use parallel computing. this can speed up the function for larger data frames}
}
\value{
a list contains the biggest value and the items
}
\description{
This function uses a brute force algorithm to solve the knapsack problem.
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n , replace = TRUE),
v = runif(n = n, 0 , 10000))
bfk <- brute_force_knapsack(x = knapsack_objects[1:8, ], W = 3500)

}
